CR changes + adding new env var for modules
